# Batching rules matrix exp