--- a/aten/src/ATen/functorch/BatchRulesLinearAlgebra.cpp
+++ b/aten/src/ATen/functorch/BatchRulesLinearAlgebra.cpp
@@ -331,11 +331,9 @@ threeOutputs linalg_lu_factor_ex_batch_rule(
 }
 
 oneOutput matrix_exp_batch_rule(const Tensor& self, c10::optional<int64_t> self_bdim) {
-  TORCH_CHECK(
-      rankWithoutBatchDim(self, self_bdim) >= 2,
-      "torch.matrix_exp: The input tensor A must have at least 2 dimensions.");
+  TORCH_CHECK(rankWithoutBatchDim(self, self_bdim) >= 2, "torch.matrix_exp: The input tensor A must have at least 2 dimensions.");
   const auto self_ = moveBatchDimToFront(self, self_bdim).contiguous();  // seems to be a bug
-  return std::make_tuple(at::linalg_matrix_exp(self_), 0);
+  return std::make_tuple(at::matrix_exp(self_), 0);
 }
 
 fourOutputs solve_ex_batch_rule(