--- a/test/functorch/test_vmap.py
+++ b/test/functorch/test_vmap.py
@@ -3701,7 +3701,6 @@ class TestVmapOperatorsOpInfo(TestCase):
         xfail('torch.ops.aten._efficient_attention_forward'),  # outputs ints
         xfail('resize_'),
         xfail('view_as_complex'),
-        xfail('matrix_exp'),
         xfail('fft.ihfft2'),
         xfail('fft.ihfftn'),
         xfail('allclose'),