Fixes #115992

### Result

The issue is fixed, where the warning no longer exist and speed improvement is observed. This is my first open source contribution so this might not be the best fix, let me know if there is anything I should change.

### Cause of the issue

It seems that there is already a batching rule implemented for `torch.linalg.matrix_exp` in [BatchRulesLinearAlgebra.cpp](https://github.com/pytorch/pytorch/blob/81cebca3d26b144ec0d4a98dcc11b0dd89af9b63/aten/src/ATen/functorch/BatchRulesLinearAlgebra.cpp#L333)
, but somehow `torch.matrix_exp` is not using that batching rule. I believe that these two calls should be exactly same in terms of functionality since one is just alias of another, so I just simply assign the batching rule for `torch.linalg.matrix_exp` to `troch.matrix_exp`. The change does seem to solve the problem. Corresponding test for matrix_exp batching rule is added.

### Validation of the change

On top of passing the `test_vmap.py`, I compare the run time of the old `torch.matrix_exp` to the new one. I try to compute the `vmap(torch.linalg.matrix_exp)(t)` and `vmap(torch.matrix_exp)(t)` for a tensor `t = torch.randn(9000, 3, 3)` and the following speed improvement was observed.

| Implementation                 | Time Before Change (seconds) | Time After Change (seconds) |
|--------------------------------|------------------------------------|-----------------------------------|
| `torch.linalg.matrix_exp`      | 0.6841                             | 0.5284                            |
| `torch.matrix_exp`             | 84.8330                            | 0.1462                            |

`torch.matrix_exp` has a very huge speed improvement, and no matter how many batches there are, the run times are all very fast. Note, no change is made for `torch.linalg.matrix_exp`, that should just be a reference.

### Additional question

After implementing the batching rule for `torch.matrix_exp`, I observed that it consistently performs about 5 times faster than `torch.linalg.matrix_exp`, despite them being functionally identical. Could there be an underlying reason for this discrepancy?

THANK YOU! ❤️ 


cc @mrshenli @pritamdamania87 @zhaojuanmao @satgera @rohan-varma @gqchen @aazzolini @osalpekar @jiayisuse @H-Huang @kwen2501 @awgu @penguinwu @fegin @XilunWu @wanchaol @fduwjj @wz337 @tianyu-l @wconstab @yf225 @jgong5 @mingfeima @XiaobingSuper @sanchitintel @ashokei @jingxu10 @mcarilli @ptrblck @leslie-fang-intel @EikanWang @zou3519 @Chillee @samdow @kshitij12345 @janeyx99 @voznesenskym @Guobing-Chen @zhuhaozhe @blzheng @wenzhe-nrv @jiayisunx @peterbell10 @ipiszy @chenyang78 @kadeng @muchulee8 @aakhundov @ColinPeppler @LucasLLC